Changed type of price field and added seeder to properties table

+++ database/migrations/2023_06_23_202525_create_properties_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('properties', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->double('price');
+            $table->decimal('price', 10, 2);
             $table->text('description');
             $table->integer('wc')->default(0);
             $table->integer('parking')->default(0);
